Eliminate duplicated localization test logic by extracting it into an helper protocol extension
